비교화면의 스크롤바가 싱크가 되지 않는 문제 수정, 비교 리스트 뷰의 스크롤바를 투명하게 만드는 스타일 없앰

+++ src/controller/MainWindowFXController.java
@@ -87,8 +87,8 @@ public class MainWindowFXController implements Initializable, CompareModeDisable
             items.add(rightEditorPane);
             editorSplitPane.setDividerPositions(0.5);
             compareViewScrollBar.valueProperty().addListener((observable, oldValue, newValue) -> {
-                leftPaneController.getCompareListView().scrollTo(newValue);
-                rightPaneController.getCompareListView().scrollTo(newValue);
+                leftPaneController.getCompareListView().scrollTo(newValue.intValue());
+                rightPaneController.getCompareListView().scrollTo(newValue.intValue());
             });
         }catch(IOException e){
             e.printStackTrace();
